Create RandomWalkers.java file
Write a program that takes two integer command-line arguments r and trials. In each of trials independent experiments, simulate a random walk until the random walker is at Manhattan distance r from the starting point. Print the average number of steps. As r increases, we expect the random walker to take more and more steps. Use this program to formulate a hypothesis as to how the average number of steps grows as a function of r.